my_fast_api/requirements.txt
fastapi==0.109.2
numpy==1.24.3
opencv_python==4.8.0.76
opencv_python_headless==4.8.1.78
Pillow==10.2.0
supervision==0.18.0
ultralytics==8.0.227
uvicorn==0.27.1
wandb==0.16.3
